+++ RingGrids/src/field.jl
@@ -296,7 +296,7 @@ Base.zero(field::F) where {F <: AbstractField} = F(zero(field.data), field.grid)
 Base.similar(field::F) where {F <: AbstractField} = F(similar(field.data), field.grid)
 
 # data with new type T but share grid
-Base.similar(field::F, ::Type{T}) where {F <: AbstractField, T} = F(similar(field.data, T), field.grid)
+Base.similar(field::F, ::Type{T}) where {F <: AbstractField, T} = nonparametric_type(F)(similar(field.data, T), field.grid)
 
 # data with same type T but new size (=new grid)
 function Base.similar(
@@ -457,11 +457,11 @@ import Base.Broadcast: BroadcastStyle, Broadcasted, DefaultArrayStyle
 # important to not have parameter T (eltype/number format) here to broadcast
 # automatically across the same field type but with different T
 # e.g. FullGaussianField{Float32} and FullGaussianField{Float64}
-struct FieldStyle{N, Grid, Column} <: Broadcast.AbstractArrayStyle{N} end
+struct FieldStyle{N, Grid} <: Broadcast.AbstractArrayStyle{N} end
 
 # define broadcast style for Field from its parameters
-Base.BroadcastStyle(::Type{F}) where {F <: Field{T, N, ArrayType, Grid}} where {T, N, ArrayType, Grid} =
-    FieldStyle{N, nonparametric_type(Grid), false}()
+Base.BroadcastStyle(::Type{F}) where {F <: AbstractField{T, N, ArrayType, Grid}} where {T, N, ArrayType, Grid} =
+    FieldStyle{N, nonparametric_type(Grid)}()
 
 # find a field within Broadcasted to reuse its grid
 find_field(bc::Base.Broadcast.Broadcasted) = find_field(bc.args)
@@ -474,51 +474,43 @@ find_field(::Any, rest) = find_field(rest)
 # allocation for broadcasting via similar, reusing grid from the first field of the broadcast arguments
 # e.g. field1 + field2 creates a new field that share the grid of field1
 # 2 .+ field1 creates a new field that share the grid of field1
-function Base.similar(bc::Broadcasted{FieldStyle{N, Grid, false}}, ::Type{T}) where {N, Grid, T}
+function Base.similar(bc::Broadcasted{FieldStyle{N, Grid}}, ::Type{T}) where {N, Grid, T}
     field = find_field(bc)
-    ArrayType_ = nonparametric_type(typeof(field.data))
-    new_data = ArrayType_{T}(undef, size(bc))
-    old_grid = field.grid
-    return Field(new_data, old_grid)
+    # parent of broadcasted arrays is used because we don't want e.g. a view or transpose as a result
+    return nonparametric_type(typeof(field))(similar(parent(field.data), T, axes(bc)), field.grid)
 end
 
 # ::Val{0} for broadcasting with 0-dimensional, ::Val{1} for broadcasting with vectors, etc
-# when there's a dimension mismatch always choose the larger dimension
-FieldStyle{N, Grid, Column}(::Val{N}) where {N, Grid, Column} = FieldStyle{N, Grid, Column}()
-FieldStyle{1, Grid, Column}(::Val{2}) where {Grid, Column} = FieldStyle{2, Grid, Column}()
-FieldStyle{1, Grid, Column}(::Val{0}) where {Grid, Column} = FieldStyle{1, Grid, Column}()
-FieldStyle{2, Grid, Column}(::Val{3}) where {Grid, Column} = FieldStyle{3, Grid, Column}()
-FieldStyle{2, Grid, Column}(::Val{1}) where {Grid, Column} = FieldStyle{2, Grid, Column}()
-FieldStyle{3, Grid, Column}(::Val{4}) where {Grid, Column} = FieldStyle{4, Grid, Column}()
-FieldStyle{3, Grid, Column}(::Val{2}) where {Grid, Column} = FieldStyle{2, Grid, Column}()
+# when there's a dimension mismatch always choose the smaller dimension
+FieldStyle{N, Grid}(::Val{N}) where {N, Grid} = FieldStyle{N, Grid}()
+FieldStyle{1, Grid}(::Val{2}) where {Grid} = FieldStyle{2, Grid}()
+FieldStyle{1, Grid}(::Val{0}) where {Grid} = FieldStyle{1, Grid}()
+FieldStyle{2, Grid}(::Val{3}) where {Grid} = FieldStyle{3, Grid}()
+FieldStyle{3, Grid}(::Val{4}) where {Grid} = FieldStyle{4, Grid}()
 
 ## GPU (same but <: GPUArrays.AbstractGPUArrayStyle)
-struct FieldGPUStyle{N, Grid, Column} <: GPUArrays.AbstractGPUArrayStyle{N} end
+struct FieldGPUStyle{N, Grid} <: GPUArrays.AbstractGPUArrayStyle{N} end
 
 # same as for FieldStyle but for constrain to ArrayType<:GPUArrays
 function Base.BroadcastStyle(
         ::Type{F}
-    ) where {F <: Field{T, N, ArrayType, Grid}} where {T, N, ArrayType <: GPUArrays.AbstractGPUArray, Grid}
-    return FieldGPUStyle{N, Grid, false}()
+    ) where {F <: AbstractField{T, N, ArrayType, Grid}} where {T, N, ArrayType <: GPUArrays.AbstractGPUArray, Grid}
+    return FieldGPUStyle{N, nonparametric_type(Grid)}()
 end
 
-function Base.similar(bc::Broadcasted{FieldGPUStyle{N, Grid, false}}, ::Type{T}) where {N, Grid, T}
+function Base.similar(bc::Broadcasted{FieldGPUStyle{N, Grid}}, ::Type{T}) where {N, Grid, T}
     field = find_field(bc)
-    ArrayType_ = nonparametric_type(typeof(field.data))
-    new_data = ArrayType_{T}(undef, size(bc))
-    old_grid = field.grid
-    return Field(new_data, old_grid)
+    # parent of broadcasted arrays is used because we don't want e.g. a view or transpose as a result
+    return nonparametric_type(typeof(field))(similar(parent(field.data), T, axes(bc)), field.grid)
 end
 
 # ::Val{0} for broadcasting with 0-dimensional, ::Val{1} for broadcasting with vectors, etc
 # when there's a dimension mismatch always choose the larger dimension
-FieldGPUStyle{N, Grid, Column}(::Val{N}) where {N, Grid, Column} = FieldGPUStyle{N, Grid, Column}()
-FieldGPUStyle{1, Grid, Column}(::Val{2}) where {Grid, Column} = FieldGPUStyle{2, Grid, Column}()
-FieldGPUStyle{1, Grid, Column}(::Val{0}) where {Grid, Column} = FieldGPUStyle{1, Grid, Column}()
-FieldGPUStyle{2, Grid, Column}(::Val{3}) where {Grid, Column} = FieldGPUStyle{3, Grid, Column}()
-FieldGPUStyle{2, Grid, Column}(::Val{1}) where {Grid, Column} = FieldGPUStyle{2, Grid, Column}()
-FieldGPUStyle{3, Grid, Column}(::Val{4}) where {Grid, Column} = FieldGPUStyle{4, Grid, Column}()
-FieldGPUStyle{3, Grid, Column}(::Val{2}) where {Grid, Column} = FieldGPUStyle{2, Grid, Column}()
+FieldGPUStyle{N, Grid}(::Val{N}) where {N, Grid} = FieldGPUStyle{N, Grid}()
+FieldGPUStyle{1, Grid}(::Val{2}) where {Grid} = FieldGPUStyle{2, Grid}()
+FieldGPUStyle{1, Grid}(::Val{0}) where {Grid} = FieldGPUStyle{2, Grid}()
+FieldGPUStyle{2, Grid}(::Val{3}) where {Grid} = FieldGPUStyle{3, Grid}()
+FieldGPUStyle{3, Grid}(::Val{4}) where {Grid} = FieldGPUStyle{4, Grid}()
 
 function KernelAbstractions.get_backend(
         field::F
@@ -526,7 +518,7 @@ function KernelAbstractions.get_backend(
     return KernelAbstractions.get_backend(field.data)
 end
 
-Adapt.parent_type(::Type{<:Field{T, N, ArrayType}}) where {T, N, ArrayType} = ArrayType
+Adapt.parent_type(::Type{<:AbstractField{T, N, ArrayType}}) where {T, N, ArrayType} = ArrayType
 Adapt.adapt_structure(to, field::AbstractField) = Adapt.adapt(to, field.data)
 
 Architectures.architecture(field::AbstractField) = architecture(field.grid)