+++ SpeedyWeather/src/output/feedback.jl
@@ -57,7 +57,7 @@ function initialize!(feedback::Feedback, clock::Clock, model::AbstractModel)
     return feedback.progress_meter = ProgressMeter.Progress(clock.n_timesteps - 1; enabled = verbose, showspeed, desc)
 end
 
-# fallback if feedback is set to nothing 
+# fallback if feedback is set to nothing
 initialize!(::Nothing, clock::Clock, model::AbstractModel) = nothing
 
 progress!(feedback::Feedback) = ProgressMeter.next!(feedback.progress_meter)
@@ -72,7 +72,7 @@ $(TYPEDSIGNATURES)
 Finalises the progress meter and the progress txt file."""
 finalize!(F::Feedback) = ProgressMeter.finish!(F.progress_meter)
 
-# fallback if feedback is set to nothing 
+# fallback if feedback is set to nothing
 finalize!(::Nothing) = nothing
 
 """$(TYPEDSIGNATURES)